Did some renaming
The "ClientIT" class is going to be more of a full-flow simulation and interaction tests more specific to the client alone will be named ClientIT
