+++ src/test/java/io/github/incplusplus/peerprocessing/SanityCheckTest.java
@@ -2,7 +2,7 @@ package io.github.incplusplus.peerprocessing;
 
 import org.junit.jupiter.api.Test;
 
-import static io.github.incplusplus.peerprocessing.ClientIT.VERBOSE_TEST_OUTPUT;
+import static io.github.incplusplus.peerprocessing.SingleSlaveIT.VERBOSE_TEST_OUTPUT;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class SanityCheckTest {
